Update integration tests
