+++ tests/integration/conftest.py
@@ -3,7 +3,9 @@
 
 import json
 
+import pytest
 import pytest_asyncio
+from juju.model import Model
 from pytest_operator.plugin import OpsTest
 
 
@@ -26,3 +28,40 @@ async def fixture_get_unit_ips(ops_test: OpsTest):
         )
 
     return get_unit_ips
+
+
+@pytest_asyncio.fixture(scope="module", name="model")
+async def fixture_model(ops_test: OpsTest) -> Model:
+    """Return the current testing juju model."""
+    assert ops_test.model
+    return ops_test.model
+
+
+@pytest.fixture(scope="module", name="external_hostname")
+def external_hostname_fixture() -> str:
+    """Return the external hostname for ingress-related tests."""
+    return "juju.test"
+
+
+@pytest.fixture(scope="module", name="traefik_app_name")
+def traefik_app_name_fixture() -> str:
+    """Return the name of the traefik application deployed for tests."""
+    return "traefik-k8s"
+
+
+@pytest.fixture(scope="module", name="prometheus_app_name")
+def prometheus_app_name_fixture() -> str:
+    """Return the name of the prometheus application deployed for tests."""
+    return "prometheus-k8s"
+
+
+@pytest.fixture(scope="module", name="loki_app_name")
+def loki_app_name_fixture() -> str:
+    """Return the name of the prometheus application deployed for tests."""
+    return "loki-k8s"
+
+
+@pytest.fixture(scope="module", name="grafana_app_name")
+def grafana_app_name_fixture() -> str:
+    """Return the name of the grafana application deployed for tests."""
+    return "grafana-k8s"
